chore: template refactoring and fix options
