+++ src/Uno.Extensions.Templates/content/unoapp-extensions/Directory.Packages.props
@@ -3,20 +3,10 @@
     <PackageVersion Include="CommunityToolkit.Mvvm" Version="8.0.0" />
     <PackageVersion Include="coverlet.collector" Version="3.2.0" />
     <PackageVersion Include="FluentAssertions" Version="6.8.0" />
-    <!--#if (msftextv6) -->
-    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="6.0.0" />
-    <!--#endif-->
-    <!--#if (msftextv7) -->
-    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="7.0.0" />
-    <!--#endif-->
+    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="$msftextlibversion$" />
     <PackageVersion Include="Microsoft.Net.Compilers.Toolset" Version="4.2.0" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
-    <!--#if (msftextv6) -->
-    <PackageVersion Include="Microsoft.Windows.Compatibility" Version="6.0.0" />
-    <!--#endif-->
-    <!--#if (msftextv7) -->
-    <PackageVersion Include="Microsoft.Windows.Compatibility" Version="7.0.0" />
-    <!--#endif-->
+    <PackageVersion Include="Microsoft.Windows.Compatibility" Version="$msftextlibversion$" />
     <PackageVersion Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.755" />
     <PackageVersion Include="Microsoft.WindowsAppSDK" Version="1.2.221209.1" />
     <PackageVersion Include="Newtonsoft.Json" Version="13.0.2" />
