+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Skia.WPF/MyExtensionsApp.Skia.WPF.csproj
@@ -16,6 +16,17 @@
 			<_Parameter2_IsLiteral>true</_Parameter2_IsLiteral>
 		</AssemblyAttribute>
 	</ItemGroup>
+	<!--#if (blank-app-template)-->
+	<ItemGroup>
+		<!--#if (use-csharp-markup)-->
+		<PackageReference Include="Uno.WinUI.Markup" />
+		<!--#endif-->
+		<PackageReference Include="Uno.WinUI.Skia.Wpf" />
+		<PackageReference Include="Uno.WinUI.RemoteControl" Condition="'$(Configuration)'=='Debug'" />
+		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
+	</ItemGroup>
+	<!--#endif-->
+	<!--#if (default-app-template)-->
 	<ItemGroup>
 		<!--#if (use-csharp-markup)-->
 		<PackageReference Include="Uno.WinUI.Markup" />
@@ -54,7 +65,7 @@
 		<!--#if (use-serilog)-->
 		<PackageReference Include="Uno.Extensions.Logging.Serilog" />
 		<!--#endif-->
-		<!--#if (!blank-app-template)-->
+		<!--#if (!use-frame-nav)-->
 		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" />
 		<PackageReference Include="Uno.Extensions.Navigation.WinUI" />
 		<!--#endif-->
@@ -72,6 +83,7 @@
 		<PackageReference Include="Uno.Extensions.Serialization.Refit" />
 		<!--#endif-->
 	</ItemGroup>
+	<!--#endif-->
 
 	<ItemGroup>
 		<ApplicationDefinition Include="Wpf\App.xaml" XamlRuntime="Wpf" />
