+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp/MainPage.xaml
@@ -7,7 +7,8 @@
 	mc:Ignorable="d"
 	Background="{ThemeResource BackgroundBrush}">
 
-	<Grid>
+	<StackPanel HorizontalAlignment="Center"
+				VerticalAlignment="Center">
 		<TextBlock Text="Hello Uno Platform" />
-	</Grid>
+	</StackPanel>
 </Page>
