+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp/MainPage.xaml.cs
@@ -1,5 +1,4 @@
 //-:cnd:noEmit
-
 namespace MyExtensionsApp;
 
 public sealed partial class MainPage : Page
@@ -8,13 +7,13 @@ public sealed partial class MainPage : Page
 	{
 //+:cnd:noEmit
 #if use-csharp-markup
-		this.Content(new Grid().Children(
-			new TextBlock()
-				.Grid(row: 0, column: 0)
-				.Text("Hello Uno Platform!")
-				.HorizontalAlignment(HorizontalAlignment.Center)
-				.VerticalAlignment(VerticalAlignment.Center)
-		));
+		this.Content(new StackPanel()
+			.VerticalAlignment(VerticalAlignment.Center)
+			.HorizontalAlignment(HorizontalAlignment.Center)
+			.Children(
+				new TextBlock()
+					.Text("Hello Uno Platform!")
+			));
 #else
 		this.InitializeComponent();
 #endif
