+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp/MyExtensionsApp.csproj
@@ -30,7 +30,7 @@
 		<!--#endif-->
 		<PackageReference Include="Uno.Themes.WinUI.Markup" />
 		<!--#endif-->
-		<!--#if (!blank-app-template)-->
+		<!--#if (default-app-template)-->
 		<!--#if (!reactive) -->
 		<PackageReference Include="CommunityToolkit.Mvvm" />
 		<!--#endif -->
