+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp/Presentation/SecondPage.xaml
@@ -13,7 +13,7 @@
 			<utu:NavigationBar.MainCommand>
 				<AppBarButton>
 					<AppBarButton.Icon>
-						<BitmapIcon UriSource="ms-appx:///Assets/Icons/back.png" />
+						<BitmapIcon UriSource="ms-appx:///MyExtensionsApp/Assets/Icons/back.png" />
 					</AppBarButton.Icon>
 				</AppBarButton>
 			</utu:NavigationBar.MainCommand>
