+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp/Styles/MaterialFontsOverride.xaml
@@ -1,8 +1,8 @@
 <ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-	<FontFamily x:Key="MaterialLightFontFamily">ms-appx:///Assets/Fonts/Material/Roboto-Light.ttf#Roboto</FontFamily>
-	<FontFamily x:Key="MaterialMediumFontFamily">ms-appx:///Assets/Fonts/Material/Roboto-Medium.ttf#Roboto</FontFamily>
-	<FontFamily x:Key="MaterialRegularFontFamily">ms-appx:///Assets/Fonts/Material/Roboto-Regular.ttf#Roboto</FontFamily>
-	
+	<FontFamily x:Key="MaterialLightFontFamily">ms-appx:///MyExtensionsApp/Assets/Fonts/Material/Roboto-Light.ttf#Roboto</FontFamily>
+	<FontFamily x:Key="MaterialMediumFontFamily">ms-appx:///MyExtensionsApp/Assets/Fonts/Material/Roboto-Medium.ttf#Roboto</FontFamily>
+	<FontFamily x:Key="MaterialRegularFontFamily">ms-appx:///MyExtensionsApp/Assets/Fonts/Material/Roboto-Regular.ttf#Roboto</FontFamily>
+
 </ResourceDictionary>
